[HACK] Changes token to i128 bit
Cassandra clusters using RandomPartitioner generate 128-bit tokens.
This is a hack to fix a parsing error when connecting to these clusters.
Note that the Token Aware load balancer still ASSUMES Murmur3 and is
will not load balance correctly if the Cassandra OR Scylla cluster are
using a different load balancing strategy.

+++ examples/compare-tokens.rs
@@ -29,7 +29,7 @@ async fn main() -> Result<()> {
             .await?;
 
         let serialized_pk = (pk,).serialized()?.into_owned();
-        let t = murmur3_token(prepared.compute_partition_key(&serialized_pk)?).value;
+        let t = murmur3_token(prepared.compute_partition_key(&serialized_pk)?).value as i64;
 
         let qt = session
             .query(format!("SELECT token(pk) FROM ks.t where pk = {}", pk), &[])
